Everything for login minus navbar
Everything that is needed for login excluding the topnavbar stuff has been done. The forms.css file has been started, the login.html is done, the login,css is one, and the login script is done. An error message will be displayed if the form is incomplete.  Even the AJAX is set up, including everything that we will need for authentication token cookies.